requirements.txt
beautifulsoup4>=4.12.3
langchain>=0.1.16
langchain_community>=0.0.34
langchain_core>=0.1.45
langchain_openai>=0.1.3
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
Requests>=2.31.0
chromadb>=0.4.24
gradio>=4.27.0

# fastapi>=0.110.2
# langserve>=0.1.0
# uvicorn>=0.29.0
